admin/requirements.txt
streamlit
pymongo
python-dotenv
pandas
altair
plotly